+++ source/runner/templates/exceptionRaisedTemplate.ts
@@ -1,0 +1,12 @@
+const quotes = "```"
+
+export default (error: Error) => `
+## Danger has errored
+
+Error: ${error.name}
+
+${quotes}sh
+${error.stack}
+${quotes}
+
+`